refactor: split claim verification workflow
